ExtensionSystem: Clarify text in "Problematic Plugin" dialog
If Qt Creator crashed at startup, we show a dialog pointing fingers to a
plugin, the next time Qt Creator is started.

Make clearer that the problem was the _last_ time Qt Creator was
started, not during the current startup.

Make clearer that the "following plugins are disabled too" refers to the
case when the user decides to temporarily disable the plugin.

Change-Id: Ie6cfed2e445b9e5c1598783474ce8a169d82bc3a
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>