Forward declare QtcProcess instead of #include
The QtcProcess is already forward declared in idevice.h.

In case of std::unique_ptr<MyClass> being a member
of a public class in a header file it's enough to
deinline constructor and destructor of a class
containing the member, and forward declare MyClass.

Amends f91a9cb8916183377dee6c708b2f4ca168710f88

Change-Id: Iacf272b62fed8827c0451c5d4c5c3429bb7106ef
Reviewed-by: Christian Stenger <christian.stenger@qt.io>