QNX: don't set LD_LIBRARY_PATH
Setting this variable causes an instance of LD_LIBRARY_PATH containing
<qtdir>/lib to be added to the environment. In the context of QNX x64,
this tricks the Linux dynamic loader into attempto to the target Qt
libraries (e.g. libQt6Core.so) instead of the host ones when running
host tools such as qmake, as the loader
will find a matching library dependency in LD_LIBRARY_PATH before it can
even reach the ld cache. This will obviously fail.

An instance of LD_LIBRARY_PATH containing the path to the Qt libraries
should not be required anyway - these explicitly passed as linker flags
to the compiler driver. Host tools rely on the builting RPATH of their
binaries.

Task-number: QTCREATORBUG-27287
Change-Id: I0527a7a77c6c76fa0229ea65695cc0d9f7d9811b
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: hjk <hjk@qt.io>