Fonts: Provide bold italic font
Linux fails to automatically mix bold and italic styles of
fonts while other OS seem to be able to handle this
appropriate.
Provide a bold italic font to not rely on the underlying
font engine.
As on it update the other fonts to the same version.
(Roman fonts version 2.038, Italic fonts version 1.058)

Fixes: QTCREATORBUG-27355
Change-Id: I68884a11dd33c57372f946dbf9ee05f04b7ebc36
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>
Reviewed-by: Eike Ziller <eike.ziller@qt.io>
Reviewed-by: David Schulz <david.schulz@qt.io>