Mimetypes v3: Improve performance
When looking up mimetypes by glob or magic, we may not add a mimetype if
it was overwritten by another provider that did not find a match.

The patch that implemented that was very inefficient, because a huge
list of "handled mimetypes" was created that needed to be checked,
every time a mimetype was determined.

Instead inform each provider about their overridden mimetypes once after
loading the mimetypes.

Amends 98b1e82d2b7000c61b7a58f36318ae8c06c248c7

Fixes: QTCREATORBUG-27319
Change-Id: I409856c272c334798d7a71ce590ff2d4f69c71a1
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>