StudioWelcome: Disable welcome page if inactive
On Windows 11 a side effect of having the welcome page
active if up to 10% CPU load (one core) in the GUI thread if the user interacts with
the GUI. If idle QDS takes 0%, but just clicking at menus can take up
to 10% CPU, if the welcome page is active.

See QTBUG-102549 and QDS-6750 for profiling data.
This seems to be related to accessibility.


Change-Id: Id098579003ef63d2d06bfb1f2afaeca15c3bbc70
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>