QbsSession: Simplify the logic
QbsSession::quit() is called only from QbsSession d'tor.
Move this method into private section.

Since QbsSession::quit() body was the only place in code where
we potentially set the m_state into State::ShuttingDown, there is no
need to check whether the m_state is set to State::ShuttingDown,
as that's impossible, since quit() can only be called once
during the lifetime of QbsSession.

Replace a call to quit() in d'tor with quit() body.

Since the ShuttingDown state was set only in d'tor, after
disconnecting from qbsProcess signals, it's not possible that
any signal handler would be called when being in ShuttingDown state.
That's why the SettingDown state is removed in this patch as
being no-op.

Change-Id: Iff77f4119d7122a9723de4a5cd0bb98f4dce8506
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>