QmlDesigner: Properly parse variant type
If a property actually is a variant we have to properly parse
the string and convert it to the correct type.
Booleans and numbers were not probably converted.
This did not create many issues, since the conversion happened
later, but it broke copy and paste and merging.
In Qt 5 this conversion seemed to happen mostly automatically.

Boolean literals have to be handled explcitly, since e.g. "10"
is technically true and can be interpreted as boolean.

Task-numbner: QDS-5944
Change-Id: I35c7cae7041667c7eac81e36a285a394263b35a4
Reviewed-by: Vikas Pachdha <vikas.pachdha@qt.io>