Warn when editing generated files
We already warn when editing a file that is not corresponding to a node
in the project and not under some project directory and not under some
project's version control.

Generated files are part of the project tree (not shown by default, if
"Hide Generated Files" is not turned off in the filter settings). So,
add another warning that the file is generated when editing a file that
is part of the project tree, but has the isGenerated flag set.

Task-number: QTCREATORBUG-27173
Change-Id: Id554d0e97bd5e033e957e7b3a863897845b7b654
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>