(Re-)Start replacing ApplicationLauncher in SimpleTargetRunner
This is a second attempt, as just replacing with QtcProcess lost
functionality. Try again with baby steps.

This here effectively reverts commit b05fc053ef08352a951baa4667b215ddc98620c
and starts with pimpl'ing SimpleTargetRunner instead. Plan is to
follow-up with dissolving the ApplicationLauncher in the pimpl,
combine combinable pieces, and finally clean up. Ideally, not much
more that a slightly customized QtcProcess should be left.

Change-Id: I69e7b01588259fe60bb76ef1ba741a183322c4de
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>