AutoTest: Fix catch2 parsing when using prefixed commands
If Catch2 is used with CATCH_CONFIG_PREFIX_ALL the relevant macros
get a prefix. This patch enhances the respective parser to be able
to handle these appropriate.
This does not really take into account whether the define is set or
not, so the parse results will only be correct if the project can
be built.

Fixes: QTCREATORBUG-27704
Change-Id: I935de752ac6106524c45c027af3e0f43673c4578
Reviewed-by: David Schulz <david.schulz@qt.io>