QtSupport: Make sure Qt 6 matches a Qt 5 version requirement
... of a wizard.
For example, the QtQuick application wizard specifies "Qt.5.6" as a
required feature, which has the implicit semantics of "Qt 5.6 or
higher". On the C++ side, a QtVersion adds its exact version x.y plus
all lower ones down to x.0, in an attempt to somehow implement
versioning semantics using tags. This means that versions 6.0 and higher
did not fulfill the 5.6 requirement anymore, which was very likely not
the intention.
We therefore extend the mechanism to major version x-1, though in the
future we should probably come up with a less absurd approach.

Fixes: QTCREATORBUG-27612
Change-Id: I41d60809a02f0caddba7e702c91c01e86ecbd4ac
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Eike Ziller <eike.ziller@qt.io>