RemoteLinux: Remove CheckForFreeDiskSpaceStep
This step was of very little use out of the box, as it required the user
to pass sensible values for e.g. the required disk space. Making it
smart enough would require non-trivial effort for very little gain, in
particular as the concept hardly makes any sense at all in conjunction
with deployment via rsync.

Fixes: QTCREATORBUG-27081
Change-Id: I586fceb59947334eba19b10fcc94ecedc661908f
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>