cdb: Use CtrlCStub for cdb.exe
This makes sure that any processes started by the debugged project in
cdb are also closed when the debugger should be stopped.

It happened to me in the past that while debugging Qt Creator in Qt
Creator and closing the debugging session tens of uic.exe were left as
zombies. And for some reason they started consuming CPU time.

taskkill /f /im uic.exe was the needed command to handle the situation.

Change-Id: I498381e793405854f7c02e9fec6b6e4718c75c25
Reviewed-by: hjk <hjk@qt.io>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>